pyproject.toml
[project]
name = "lapa-ng"
version = "0.1.0"
description = "LAPA allows for converting digitised early modern Dutch theatre plays into (presumed) phonetic script (SAMPA). To achieve this, a ruleset has been created that codifies the transliteration to SAMPA. This codebase contains parsers for the rule sets (xls format), parsers for the digitised texts (naf xml) and logic to perform counts and correlations."
authors = [
    {name = "Fieke Smitskamp", email = "fieke.smitskamp@gmail.com"},
    {name = "Ruben"},
    {name = "Kaj Siebert", email = "kaj@k-si.com"},
]
license = "CC-BY-4.0"
readme = "README.md"
requires-python = ">3.10,<3.13"
dependencies = [
    "click (>=8.1.8,<9.0.0)",
    "xlrd (>=2.0.1,<3.0.0)",
    "openpyxl (>=3.1.2,<4.0.0)",
    "lxml (>=5.4.0,<6.0.0)",
    "xlwt (>=1.3.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "cachetools (>=5.5.2,<6.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "plotly (>=6.0.1,<7.0.0)"
]

[tool.poetry]
packages = [
    { include = "lapa_classic" },
    { include = "lapa_ng" }
]

[project.scripts]
lapa-ng = "lapa_ng._cli:cli"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
poetry-plugin-shell = ">=1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.3.5"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"
coverage = "^7.8.0"

